Project with the cifar10 dataset.
On english
This is my first project using Tensorflow, and I'm happy to share the results! In the file Main.py A code is presented that checks the effectiveness of a model trained on data from the CIFA-10 set. The model is saved in the my_model.h5 file, and although the training took place in the same file (the code for training the model is in the comments), the results are impressive!

My model does an excellent job of recognizing images from CIFAR-10, achieving high accuracy thanks to training in 30 epochs. This experience was an important step towards a deep understanding of the work of neural networks and their application on real data. Now that the model is ready, it can be used for efficient image classification, which opens up a lot of opportunities for further improvement and testing of various architectures. I don't know how big or small this is a step in my development, but I'm happy with myself!

На русском
Это мой первый проект с использованием Tensorflow, и я с удовольствием делюсь результатами! В файле Main.py представлен код, который проверяет эффективность модели, обученной на данных из набора CIFA-10. Модель сохранена в файле my_model.h5, и, хотя обучение происходило в том же файле (код для тренировки модели находится в комментариях), результаты впечатляют!

Моя модель отлично справляется с распознаванием изображений из CIFAR-10, достигнув высокой точности благодаря обучению в 30 эпох. Этот опыт стал важным шагом на пути к глубокому пониманию работы нейросетей и их применения на реальных данных. Теперь, когда модель готова, её можно использовать для эффективной классификации изображений, что открывает массу возможностей для дальнейшего улучшения и тестирования различных архитектур. Не знаю насколько большой или маленький это шаг в моем развитии, но я доволен собой!